[NOISSUE] Remove hardcoded admin account with support for multiple admins
[Problem]

Admin account was hard-coded for a single user. This if condition will not scale when deploying this for multiple jamaats

[Solution]

Add a new responsibility as 'A' (Admin) under the 'resp' column in the family table schema. Logging in with an email that where resp includes 'A'  gives admin privileges.

Create an admin-only account with thaali #0

Replace families with anonymous variants for privacy reasons. (These will still be visible if someone digs deep into the git history)

Modified the deploy.pl file to be used by other jamaats and their mysql configurations.

[Testing]

gulp serve on localhost. See Screenshot in PR comments.

Tested the deploy.pl changes by  deploying and testing on http://fmbvancouver.dreamhosters.com/

[Issue]

NOISSUE